module_17/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.47
python-slugify==8.0.1